refactor: split the main scripts to improve folder organization
