+++ js/builders/calculator.js
@@ -1,11 +1,14 @@
 // Dados para criação da calculadora
-import { mathData, themeData } from "./data.js"
+import { buttonDataList } from "../data/buttons.js"
 
 // Eventos de tratamento de botões da calculadora
-import { handleClickButton, handleSwitchThemeColor } from "./buttonEvents.js"
+import { handleClickButton } from "../events/buttons.js"
 
 // Eventos de entrada de teclado da calculadora
-import { numberMaskInput, handleKeyDown } from "./inputEvents.js"
+import { handleKeyDown } from "../events/inputs.js"
+
+// Máscara de input de número da calculadora
+import { numberMaskInput } from "../utils/masks/numbers.js"
 
 function buildOutput() {
   const container = document.createElement("div")
@@ -29,30 +32,19 @@ function buildInput() {
   const container = document.createElement("ul")
   container.classList.add("input")
 
-  mathData.forEach(list => {
+  buttonDataList.forEach(list => {
     const row = document.createElement("li")
     row.classList.add("row")
     
     list.forEach(item => {
-      const { type, value, label, icon } = item
+      const { type, value, label } = item
       const button = document.createElement("button")
 
       button.classList.add("button", `--${type}`)
       button.setAttribute("type", "button")
       button.setAttribute("data-value", value)
       button.addEventListener('click', handleClickButton)
-
-      if (icon) {
-        const svg = document.createElement("img")
-        svg.classList.add("icon")
-        svg.setAttribute("title", label)
-        svg.setAttribute("src", icon)
-
-        button.appendChild(svg)
-      } else {
-        button.innerText = label
-      }
-
+      button.innerHTML = label
       row.appendChild(button)
     })
 
@@ -62,29 +54,8 @@ function buildInput() {
   return container
 }
 
-export function buildThemeButton() {
-  const container = document.createElement("button")
-
-  container.id = "theme"
-  container.setAttribute("type", "button")
-  container.addEventListener("click", handleSwitchThemeColor)
-
-  themeData.forEach(item => {
-    const { theme, isVisible, alt, src } = item
-    const icon = document.createElement("img")
-
-    icon.classList.add("icon", `--${theme}`)
-    icon.setAttribute("alt", alt)
-    icon.setAttribute("src", src)
-
-    container.appendChild(icon)
-  })
-
-  return container
-}
-
 export function buildCalculator() {
-  const container = document.createElement("section")
+  const container = document.createElement("main")
   const output = buildOutput()
   const input = buildInput()
