+++ js/events/inputs.js
@@ -1,0 +1,13 @@
+import { handleClearContent, handleSolveContent } from "../events/buttons.js"
+
+export function handleKeyDown(event) {
+  const { key } = event
+
+  if (key == "=" || key == "Enter") {
+    return handleSolveContent()
+  }
+
+  if (key == "Escape") {
+    return handleClearContent()
+  }
+}
