+++ js/utils/masks/numbers.js
@@ -1,0 +1,5 @@
+export function numberMaskInput(event) {
+  const display = document.getElementById("display")
+  const { target: { value }} = event
+  display.value = value.replace(/^([\/\*])|([^\d*.*()*\+\-\/\*])/gi, "")
+}
